+++ bindings/rust/evmc-declare-tests/src/lib.rs
@@ -22,7 +22,7 @@ impl EvmcVm for FooVM {
         _revision: evmc_sys::evmc_revision,
         _code: &[u8],
         _message: &ExecutionMessage,
-        _context: &mut ExecutionContext,
+        _context: Option<&mut ExecutionContext>,
     ) -> ExecutionResult {
         ExecutionResult::success(1337, None)
     }